-Seperated GUI and Converter class -Included additional AIs and set Promi to the default AI -Fixed problems due to symlinks when updating from 5.7 to 5.7.2 -Added No Snow option (Beach for Ice beach, Shallows for Ice Shallows, dlc_rock for Land Ice)
