+++ fixes/tricklebuildingfix.h
@@ -1,0 +1,12 @@
+#ifndef TRICKLEBUILDINGFIX_H
+#define TRICKLEBUILDINGFIX_H
+#include "wololo/datPatch.h"
+
+namespace wololo {
+
+extern DatPatch trickleBuildingFix;
+int getNewGraphicId(genie::DatFile *aocDat, int i);
+
+}
+
+#endif // TRICKLEBUILDINGFIX_H
