+++ mainwindow.h
@@ -7,6 +7,9 @@
 
 #include <set>
 #include <regex>
+#include "wkgui.h"
+#include "wksettings.h"
+#include "wkconverter.h"
 
 #include <boost/filesystem.hpp>
 #include "genie/dat/DatFile.h"
@@ -26,7 +29,7 @@ namespace Ui {
 class MainWindow;
 }
 
-class MainWindow : public QMainWindow
+class MainWindow : public QMainWindow, public WKGui
 {
     Q_OBJECT
 
@@ -34,55 +37,43 @@ public:
 	explicit MainWindow(QWidget *parent = 0);
     ~MainWindow();
 
-public slots:
-	int run();
-
 private:
 
     std::string steamPath;
     fs::path HDPath;
     fs::path outPath;
-    std::set<int> aocSlpFiles;
-    std::map<int, fs::path> slpFiles;
-    std::map<int, fs::path> wavFiles;
-    std::map<std::string,fs::path> newTerrainFiles;
-    std::vector<std::pair<int,std::string>> rmsCodeStrings;
     std::string version = "5.7";
-    std::string hash1;
-    std::string hash2;
-    std::string patchNumber;
     std::map<int, std::tuple<std::string,std::string, std::string, int, std::string>> dataModList;
     std::string language = "en";
     std::map<std::string, std::string> translation;
-    std::map<int, std::string> gameTranslation;
-    bool secondAttempt = false;
     bool allowRun = true;
 
-    std::set<char> civLetters;
     QProgressBar* bar = NULL;
-    int dlcLevel = 0;
+    int dlcLevel;
     int patch = -1;
     std::string modName;
     std::ofstream logFile;
 
-    fs::path nfzUpOutPath;
-    fs::path nfzVooblyOutPath;
     fs::path vooblyDir;
     fs::path upDir;
     fs::path installDir;
+    fs::path nfzUpOutPath;
+    fs::path nfzVooblyOutPath;
     std::string baseModName = "WololoKingdoms";
     fs::path resourceDir;
 
-    enum TerrainType {
-        None,
-        WaterTerrain,
-        FixedTerrain,
-        LandTerrain,
-        ForestTerrain,
-        UnbuildableTerrain
-    };
-
 	Ui::MainWindow *ui;
+
+    void log(std::string logMessage);
+    void setInfo(std::string info);
+    void createDialog(std::string info);
+    void createDialog(std::string info, std::string title);
+    void setProgress(int i);
+    void increaseProgress(int i);
+    std::string translate(std::string line);
+
+    void runConverter();
+
     int initialize();
     void setInstallDirectory(std::string directory);
     void changeLanguage();
@@ -93,40 +84,7 @@ private:
     void writeSettings();
 	void changeModPatch();
     void callExternalExe(std::wstring exe);
-	void updateUI();
-    void copyHDMaps(fs::path inputDir, fs::path outputDir, bool replace = false);
-    bool usesMultipleWaterTerrains(std::string& map, std::map<int,bool>& terrainsUsed);
-    bool isTerrainUsed(int terrain, std::map<int,bool>& terrainsUsed, std::string& map, std::map<int,std::regex>& patterns);
-    void upgradeTrees(int usedTerrain, int oldTerrain, std::string& map);
-    void createZRmap(std::map<std::string,fs::path>& terrainOverrides, fs::path outputDir, std::string mapName);
-	void terrainSwap(genie::DatFile *hdDat, genie::DatFile *aocDat, int tNew, int tOld, int slpID);
-    void recCopy(fs::path const &src, fs::path const &dst, bool skip = false, bool force = false);
-    void indexDrsFiles(fs::path const &src, bool expansionFiles = true, bool terrainFiles = false);
-    void copyHistoryFiles(fs::path inputDir, fs::path outputDir);
-    std::pair<int,std::string> getTextLine(std::string line);
-	void convertLanguageFile(std::ifstream *in, std::ofstream *iniOut, genie::LangFile *dllOut, bool generateLangDll, std::map<int, std::string> *langReplacement);
-    bool createLanguageFile(fs::path languageIniPath, fs::path patchFolder);
-    void loadModdedStrings(std::string moddedStringsFile, std::map<int, std::string>& langReplacement);
-    bool openLanguageDll(genie::LangFile *langDll, fs::path langDllPath, fs::path langDllFile);
-    bool saveLanguageDll(genie::LangFile *langDll, fs::path langDllFile);
-	void makeDrs(std::ofstream *out);
-    void uglyHudHack(fs::path);
-    void copyCivIntroSounds(fs::path inputDir, fs::path outputDir);
-    void copyWallFiles(fs::path inputDir);
-	std::string tolower(std::string line);
-	void createMusicPlaylist(std::string inputDir, std::string const outputDir);
-	void transferHdDatElements(genie::DatFile *hdDat, genie::DatFile *aocDat);
-    void adjustArchitectureFlags(genie::DatFile *aocDat, std::string flagFilename);
-	void patchArchitectures(genie::DatFile *aocDat);
-    bool checkGraphics(genie::DatFile *aocDat, short graphicID, std::vector<int> checkedGraphics);
-    void replaceGraphic(genie::DatFile *aocDat, short* graphicID, short compareID, short c, std::map<short,short>& replacedGraphics, bool civGroups = false);
-    short duplicateGraphic(genie::DatFile *aocDat, std::map<short,short>& replacedGraphics, std::vector<short> duplicatedGraphics, short graphicID, short compareID, short offset, bool civGroups = false);
-    bool identifyHotkeyFile(fs::path directory, fs::path& maxHki, fs::path& lastEditedHki);
-    void copyHotkeyFile(fs::path maxHki, fs::path lastEditedHki, fs::path dst);
-    void removeWkHotkeys();
-	void hotkeySetup();
-    void symlinkSetup(fs::path oldDir, fs::path newDir, fs::path xmlIn, fs::path xmlOut, bool dataMod = false);
-	bool copyData(QIODevice &inFile, QIODevice &outFile);
+    void updateUI();
 };
 
 #endif // MAINWINDOW_H
