+++ resources/en.txt
@@ -28,7 +28,7 @@ workingDll=<dll> patched.
 workingError=Error, trying again
 workingUP=Copying the files for UserPatch...
 workingDone=Done!
-dialogNoDll=Couldn't read/write the language_x1_p1.dll file!\nYou may need to run the installer a few times until it works, as creating this file can be unstable.\nYou can still play via Voobly without it, but for offline play the language_x1_p1.dll file is required.
+dialogNoDll=Couldn't read/write the language_x1_p1.dll file! A generic language_x1_p1.dll file was copied instead, but some strings may be inaccurate or outdated.\nYou can try running the installer again.
 dialogExe=Conversion complete! The WololoKingdoms mod is now available as a separate <exe>.exe in the age2_x1 folder.
 dialogBoth=Conversion complete! The WololoKingdoms mod is now available as a Voobly Mod and as a separate <exe>.exe in the age2_x1 folder.
 dialogDone=Conversion complete! The WololoKingdoms mod is now available as a Voobly Mod.
@@ -43,6 +43,7 @@ useVoobly=Voobly Installation
 useBoth=Both
 useGrid=Grid
 useWalls=Short Walls
+useNoSnow=No Snow
 useMonks=Regional Monks
 installLabel=Install Directory
 copyMaps=Copy Real World and Special Maps
@@ -59,38 +60,3 @@ noSteamApi=Steam Api could not be initialized! Steam needs to be running and you
 noSteamInstallation=There is no installation of Age2HD detected. It needs to be installed to run this converter.
 noAoC=No installation of Age of Conquerors detected.\nYou need to have either the CD version or the compatibility patch installed. Make sure you can run regular AoC on Voobly before running this installer.
 errorTitle=Error!
-5045=Food Workshop (Convertable)
-5046=Wood Workshop (Convertable)
-5047=Stone Workshop (Convertable)
-5048=Gold Workshop (Convertable)
-5049=Food Workshop
-5050=Wood Workshop
-5051=Stone Workshop
-5052=Gold Workshop
-5053=Resource Workshop (Convertable)
-5054=Resource Workshop
-7440=Uprooting
-8440=Uprooting (Mangonels and Onagers cut down trees)
-17440=Uprooting
-28440=Research <b>Uprooting<b> (<cost>) \\n Mangonels and Onagers can cut down trees
-10622=Dirt 1
-10626=Mangrove Forest
-10642=Mangrove Shallows
-10648=Desert
-10618=Water, Shoreless
-10619=Baobab Forest
-10668=Road, Snow
-10677=Cracked Earth
-10679=Acacia Forest
-10707=KotH (Road, Unbuildable)
-10716=Conditions & Effects
-19031=Camel, Battle Elephant, Tarkan
-19032=See Dock Menu for Genitour hotkey
-19033=Use the Camel hotkey for Battle Elephants
-19048=See Dock Menu for Siege Tower hotkey
-19053=Longboat, Caravel, Turtle Ship
-19072=Siege Tower, Genitour (previously Turtle Ships)
-19075=Use the Wonder hotkey for Feitorias
-19209=Palisade Gate, Feitoria (previously Wonder)
-19212=Use the Wonder hotkey for Palisade Gates
-30195=Relics - Win by being the first player or team to destroy all enemies in military conquest or control all Relics. Once the relics are captured, the game will end immediately with no countdown. ZR@CtR maps should be used with this condition.
